$redeem Finalized

+++ KaguyaProjectV2/KaguyaBot/DataStorage/DbData/Models/SupporterKey.cs
@@ -13,7 +13,7 @@ namespace KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Models
         [Column(Name = "LengthInDays"), NotNull]
         public int Length { get; set; }
         [Column(Name = "ActivatedByUser"), Nullable]
-        public uint UserId { get; set; }
+        public ulong UserId { get; set; }
         /// <summary>
         /// If this key is tied to a user, the time (as OADate) when the key expires.
         /// </summary>
